STY: Move read_raw_eyelinkhelper funcs to _utils module
To clena up the module with read_raw_eyelink
